Fixing bug when a member tried to delete himslef

+++ src/Controller/MembersController.php
@@ -7,6 +7,7 @@ use App\Form\MemberType;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -22,7 +23,7 @@ final class MembersController extends AbstractController
 
         return $this->render('members/index.html.twig', [
             'controller_name' => 'MembersController',
-            'members' => $members
+            'members' => $members,
         ]);
     }
 
@@ -30,7 +31,22 @@ final class MembersController extends AbstractController
     public function edit(int $id, Request $request, EntityManagerInterface $em): Response
     {
 
+        /** @var Users $user */
+        $user = $this->getUser();
         $member = $em->getRepository(Users::class)->find($id);
+
+        if (!$member) {
+            throw $this->createNotFoundException('Membre ' . $id . ' non trouvé.');
+        }
+
+        if (in_array('ROLE_ADMIN', $member->getRoles(), true)) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas modifier un administrateur.');
+        }
+
+        if ($user->getId() !== $member->getId() && (!in_array('ROLE_MANAGER', $user->getRoles(), true) || $member->getRoles() !== ['ROLE_USER'])) {
+            throw $this->createAccessDeniedException('Vous n\'avez pas les droits pour modifer ce membre.');
+        }
+
         $form = $this->createForm(MemberType::class, $member);
         $form->handleRequest($request);
 
@@ -46,15 +62,35 @@ final class MembersController extends AbstractController
     }
 
     #[Route('/delete/{id}', name: 'delete')]
-    public function delete(int $id, EntityManagerInterface $em): Response
+    public function delete(int $id, EntityManagerInterface $em, Security $security): Response
     {
+        /** @var Users $user */
+        $user = $this->getUser();
+
         $member = $em->getRepository(Users::class)->find($id);
 
-        if(!$member) {
-            throw $this->createNotFoundException('Membre' . $id . 'non trouvé.');
+        if (!$member) {
+            throw $this->createNotFoundException('Membre ' . $id . ' non trouvé.');
         }
+
+        if (in_array('ROLE_ADMIN', $member->getRoles(), true)) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas supprimer un administrateur.');
+        }
+
+        if ($user->getId() !== $member->getId() && (!in_array('ROLE_MANAGER', $user->getRoles(), true) || $member->getRoles() !== ['ROLE_USER'])) {
+            throw $this->createAccessDeniedException('Vous n\'avez pas les droits pour supprimer ce membre.');
+        }
+
+        if ($user->getId() === $member->getId()) {
+            $security->logout(false);
+        }
+
         $em->remove($member);
         $em->flush();
+
+
+
         return $this->redirectToRoute('app_members_index');
     }
+
 }
